Новая версия аддона переписанная по стандартам cs-cart.
