--- README.md
@@ -1,1 +1,0 @@
-it's module for cs-cart cms
