+++ unitpay-addon/app/addons/unitpay/addon.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0"?>
+<addon scheme="3.0" edition_type="ROOT,ULT:VENDOR">
+    <id>unitpay</id>
+    <version>3.0.0</version>
+    <default_language>ru</default_language>
+    <priority>100</priority>
+    <status>active</status>
+    <auto_install>MULTIVENDOR,ULTIMATE</auto_install>
+    <compatibility>
+    </compatibility>
+    <queries>
+        <item for="install">
+            INSERT INTO `cscart_language_values` (`lang_code`, `name`, `value`) VALUES ('ru', 'unitpay_domain', 'Домен');
+        </item>
+        <item for="install">
+            INSERT INTO `cscart_language_values` (`lang_code`, `name`, `value`) VALUES ('en', 'unitpay_domain', 'Domain');
+        </item>
+        <item for="install">
+            INSERT INTO `cscart_language_values` (`lang_code`, `name`, `value`) VALUES ('ru', 'unitpay_public_key', 'Публичный ключ');
+        </item>
+        <item for="install">
+            INSERT INTO `cscart_language_values` (`lang_code`, `name`, `value`) VALUES ('en', 'unitpay_public_key', 'Public key');
+        </item>
+        <item for="install">
+            INSERT INTO `cscart_language_values` (`lang_code`, `name`, `value`) VALUES ('ru', 'unitpay_secret_key', 'Секретный ключ');
+        </item>
+        <item for="install">
+            INSERT INTO `cscart_language_values` (`lang_code`, `name`, `value`) VALUES ('en', 'unitpay_secret_key', 'Secret key');
+        </item>
+        <item for="install">
+            INSERT INTO `cscart_payment_processors` (`processor`, `processor_script`, `processor_template`, `admin_template`, `callback`, `type`, `addon`) VALUES ('unitpay', 'unitpay.php', 'views/orders/components/payments/cc_outside.tpl', 'unitpay.tpl', 'N', 'P', 'unitpay');
+        </item>
+        <item for="uninstall">
+            DELETE FROM `cscart_language_values` WHERE `lang_code` = 'ru' AND `name` = 'unitpay_domain';
+        </item>
+        <item for="uninstall">
+            DELETE FROM `cscart_language_values` WHERE `lang_code` = 'en' AND `name` = 'unitpay_domain';
+        </item>
+        <item for="uninstall">
+            DELETE FROM `cscart_language_values` WHERE `lang_code` = 'ru' AND `name` = 'unitpay_public_key';
+        </item>
+        <item for="uninstall">
+            DELETE FROM `cscart_language_values` WHERE `lang_code` = 'en' AND `name` = 'unitpay_public_key';
+        </item>
+        <item for="uninstall">
+            DELETE FROM `cscart_language_values` WHERE `lang_code` = 'ru' AND `name` = 'unitpay_secret_key';
+        </item>
+        <item for="uninstall">
+            DELETE FROM `cscart_language_values` WHERE `lang_code` = 'en' AND `name` = 'unitpay_secret_key';
+        </item>
+        <item for="uninstall">
+            DELETE FROM `cscart_payment_processors` WHERE `processor` = 'unitpay' AND `processor_script` = 'unitpay.php';
+        </item>
+    </queries>
+</addon>
