+++ unitpay-addon/app/addons/unitpay/schemas/routes/objects.post.php
@@ -1,0 +1,10 @@
+<?php
+
+defined('BOOTSTRAP') or die('Access denied');
+
+$schema['/unitpay_callback'] = [
+    'dispatch' => 'payment_notification.notify',
+    'payment'  => 'unitpay',
+];
+
+return $schema;
