--- unitpay/app/payments/unitpay.php
@@ -1,33 +1,0 @@
-<?php
-
-$sum = $order_info['total'];
-$account = $order_id;
-$desc = 'Заказ #' . $order_id;
-
-$payment_url = 'https://' . $processor_data['processor_params']['unitpay_domain'] . '/pay/' . $processor_data['processor_params']['unitpay_public_key'];
-$secret_key = $processor_data['processor_params']['unitpay_secret_key'];
-$signature = hash('sha256', join('{up}', array(
-    $account,
-    $desc,
-    $sum,
-    $secret_key
-)));
-
-$data = array(
-    'sum' => $sum,
-    'account'   =>  $account,
-    'desc'  =>  $desc,
-    'signature' => $signature
-);
-
-if (isset($order_info['email'])) {
-    $data['customerEmail'] = $order_info['email'];
-}
-
-if (isset($order_info['phone'])) {
-    $data['customerPhone'] = preg_replace('/\D/', '', $order_info['phone']);
-}
-
-fn_change_order_status($order_id, 'O');
-
-fn_create_payment_form($payment_url, $data, 'Unitpay', false);
